Moved shared interfaces and functions to utils
