+++ src/components/mentor/MentorBookingModal.tsx
@@ -12,7 +12,7 @@ import {
   } from '@mui/material';
   import { SlotInfo } from 'react-big-calendar';
   import Modal from '../shared/Modal';
-  import { CalendarEvent, isCalendarEvent } from '../student/StudentBooking';
+  import { CalendarEvent, isCalendarEvent } from '../../utils/CalendarEvent';
   import {
     AccessTime,
     Computer,
@@ -22,7 +22,7 @@ import {
     PersonAdd,
     PhotoCameraFront,
   } from '@mui/icons-material';
-  import { SlotSessions } from '../student/StudentBooking';
+  import { SlotSessions } from '../../utils/Session';
 import { useState } from 'react';
   
   interface EventModalProps {
