+++ src/utils/CalendarEvent.ts
@@ -1,5 +1,7 @@
 
 import {ReactNode } from 'react';
+import { SlotSessions } from './Session';
+import {SlotInfo } from 'react-big-calendar';
 
 export interface CalendarEvent {
     id: string;
@@ -11,3 +13,10 @@ export interface CalendarEvent {
     mentor?: string;
     student?: string;
   }
+
+  export const isCalendarEvent = (
+    event: CalendarEvent | SlotInfo | null | SlotSessions
+  ): event is CalendarEvent => {
+    return event ? 'id' in event : false;
+  };
+
